Changes to session expiring:
Cleared the expire_at session variable in logout_keeping_session
Do we want a remember_me_token? that would allow people to be logged in for
long periods of time like 2 weeks. Expiring sessions due to inactivity
seems the opposite of this. The code for remember_me is in the application
controller but commented out for now.